prepared for spring 4.0.0

+++ src/main/kotlin/vitos/local/keycloak_kotlin/models/BruteForceStatus.kt
@@ -20,6 +20,6 @@ data class BruteForceStatus(
     @JsonProperty("numFailures") var numFailures: Int? = null,
     @JsonProperty("disabled") var disabled: Boolean? = null,
     @JsonProperty("lastIPFailure") var lastIPFailure: String? = null,
-    @JsonProperty("lastFailure") var lastFailure: Int? = null,
+    @JsonProperty("lastFailure") var lastFailure: Long? = null
 
-    )
+)
